Reorganize StoryManager as Interface
